move Sim2hState into a module
